requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0

SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

sentence-transformers==2.7.0
google-genai
faiss-cpu==1.9.0
numpy==1.26.3

requests==2.31.0